requirements.txt
streamlit>=1.36.0
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
plotly>=5.22.0
joblib>=1.3.2
tensorflow-cpu==2.20.0
